requirements.txt
langchain
langchain-google-genai
python-dotenv
pydantic
pymongo
bcrypt
